Check if the cachedir exists before we run find on it.

+++ cobbler/action_sync.py
@@ -213,10 +213,11 @@ class BootSync:
         utils.rmtree_contents(self.rendered_dir,logger=self.logger)
 
     def clean_link_cache(self):
-        for dirtree in [self.bootloc, self.settings.webdir]:
-            cachedir = '%s/.link_cache'%dirtree
-            cmd = "find %s -maxdepth 1 -type f -links 1 -exec rm -f '{}' ';'"%cachedir
-            utils.subprocess_call(self.logger,cmd)
+        for dirtree in [os.path.join(self.bootloc,'images'), self.settings.webdir]:
+            cachedir = os.path.join(dirtree,'.link_cache')
+            if os.path.isdir(cachedir):
+                cmd = "find %s -maxdepth 1 -type f -links 1 -exec rm -f '{}' ';'"%cachedir
+                utils.subprocess_call(self.logger,cmd)
 
     def rsync_gen(self):
         """
